Implement analytic page modal
Added an modal on the dashboard for the analytic page that dynamically
changes based on the bus id clicked on the enquiry table. This should be
better for user accessability as they are no longer routed to a new page
upon clicking on the ids. Instead, a modal with static backdrop shows up
containing the AnalyticPage component.